backend/pyproject.toml
[tool.poetry]
name = "bevflix"
version = "0.1.0"
description = "netflix clone"
authors = ["esilean <le.bevilaqua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
pydantic = {extras = ["email"], version = "^1.10.8"}
motor = "^3.1.2"
psutil = "^5.9.5"
python-decouple = "^3.8"
pymongo = "^4.3.3"
pyyaml = "^6.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
prometheus-fastapi-instrumentator = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
